Gen AI Projects/ai_resume_parser/requirements.txt
streamlit
crewai
python-docx
pymupdf
crewai-tools